+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## CHANGELOG
 
+### 2.2.0
+
+* Implement new option `--with-notice-file`
+* Added to find British style file name `LICENCE` with run '--with-license-file'
+
 ### 2.1.1
 
 * Suppress errors when opening license files